Refactor the generation of possible moves
As the logic for capturing the opponent's pieces is similar to the
generation of possible moves for a player, the method was refactored and
now it is more general. It returns a list of all pieces that are
captured by the player. If that list is not empty, then the cell is a
possible move.